strange compilation issue when adding a member in Transaction.

+++ cacheLight/mainwindow.h
@@ -27,7 +27,7 @@ private:
 	Ui::MainWindow *ui;
 	QDate m_date;
 	int m_d0 = 0;
-	int m_lastBal = 0;
+	double m_lastBal = 0.0;
 	int m_ipb = 0;
 	// account to read predicted transactions from
 	Account m_account;
